requirements.txt
google-cloud-bigquery>=3.11.4
pandas>=2.0.0 